Some walls for Downtown, tried to fix main menu

+++ Assets/Scripts/MainMenuCameraScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MainMenuCameraScript : MonoBehaviour
 {
-    public int rotationSpeed;
+    public float rotationSpeed;
 
     // Start is called before the first frame update
     void Start()
